Corrigido erros de compilaçao para hash. Corrigido erro de redefinicao classe NoHash. Mudança total de tipo das metricas

+++ src/algoritmos/hash/duploHash.cpp
@@ -108,6 +108,6 @@ int duploHash::getPrimo(){
 
 }
 
-int duploHash::getNumColisoes(){
+unsigned long long int duploHash::getNumColisoes(){
     return this->numColisoes;
 }
