+++ src/algoritmos/hash/hashEncadSeparado.h
@@ -16,6 +16,6 @@ class HashEncadSeparado{
         void removeItem(int chave);
         int funcaoHash(int chave);
         void imprimirHash();
-        int contabilizarColisoes();
+        unsigned long long int contabilizarColisoes();
 
 };
