+++ src/algoritmos/ordenacao/HeapSort.h
@@ -1,7 +1,7 @@
 #ifndef HEAPSORT_H
 #define HEAPSORT_H
 
-void heapify(int *id, int tam, int indice, int *trocas, int *comparacoes);
-void heapSort(int *id, int tam, int *trocas, int *comparacoes);
+void heapify(int *id, int tam, int indice, unsigned long long int *trocas, unsigned long long int *comparacoes);
+void heapSort(int *id, int tam, unsigned long long int *trocas, unsigned long long int *comparacoes);
 
 #endif // HEAPSORT_H
