+++ src/algoritmos/ordenacao/quickSortMediano.h
@@ -4,8 +4,8 @@
 
 void gerarVetor(int* v, int k, int* aux);
 
-int medianadek(int *vetor, int k, int *trocas, int *comparacoes);
+int medianadek(int *vetor, int k, unsigned long long int *trocas, unsigned long long int *comparacoes);
 
-void quickSortMediano(int *vetor, int inicio, int fim, int k, int *trocas, int *comparacoes);
+void quickSortMediano(int *vetor, int inicio, int fim, int k, unsigned long long int *trocas, unsigned long long int *comparacoes);
 
 #endif // QUICKSORTMEDIANO_H
